+++ RRGame/Classes/Scenes/RRGameScene/RRGameScene.m
@@ -19,12 +19,12 @@
 #pragma mark UDGameScene
 
 
-+ (id)sceneWithGameMode:(RRGameMode)gameMode numberOfPlayers:(NSUInteger)numberOfPlayers firstPlayerColor:(RRPlayerColor)playerColor {
++ (id)sceneWithGameMode:(RRGameMode)gameMode numberOfPlayers:(uint)numberOfPlayers firstPlayerColor:(RRPlayerColor)playerColor {
     return [[[self alloc] initWithGameMode:gameMode numberOfPlayers:numberOfPlayers playerColor:playerColor] autorelease];
 }
 
 
-- (id)initWithGameMode:(RRGameMode)gameMode numberOfPlayers:(NSUInteger)numberOfPlayers playerColor:(RRPlayerColor)playerColor {
+- (id)initWithGameMode:(RRGameMode)gameMode numberOfPlayers:(uint)numberOfPlayers playerColor:(RRPlayerColor)playerColor {
     if( (self = [self init]) ){
         _numberOfPlayers = numberOfPlayers;
         RRGameLayer *gameLayer;
