+++ RRGame/Classes/Scenes/RRPickColorScene/RRPickColorScene.h
@@ -11,7 +11,7 @@
 
 @interface RRPickColorScene : CCScene
 
-+ (id)sceneWithNumberOfPlayers:(NSUInteger)numberOfPlayers;
-- (id)initWithNumberOfPlayers:(NSUInteger)numberOfPlayers;
++ (id)sceneWithNumberOfPlayers:(uint)numberOfPlayers;
+- (id)initWithNumberOfPlayers:(uint)numberOfPlayers;
     
 @end
